state of job added and used

+++ src/include/cellBE/GeneralFilterSerializer.h
@@ -30,7 +30,7 @@ public:
     , AbstractFilterSerializer **serializer
     , M4D::CellBE::NetStream &s)
   {
-	  uint8 filterID;
+	  uint16 filterID;
 	  s >> filterID;
 
     uint16 id;
